pyproject.toml
[tool.poetry]
name = "collabland-action-fastapi"
version = "0.1.0"
description = "A FastAPI Template for Collab.Land Actions"
authors = ["Abridged, Inc.","Gita Alekhya Paul <54375111+gitaalekhyapaul@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "collabland_action_fastapi"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.93.0"
uvicorn = "^0.21.0"
discord-py = "^2.2.2"
python-dotenv = "^1.0.0"
requests = "^2.28.2"
base58 = "^2.1.1"
pynacl = "^1.5.0"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = 'collabland_action_fastapi.main:start'